Moved player class under entities package
