+++ src/com/rech/rpg/FileManager.java
@@ -5,6 +5,8 @@ import java.io.File;  // Import the File class
 import java.io.FileNotFoundException;  // Import this class to handle errors
 import java.util.Scanner; // Import the Scanner class to read text files
 
+import com.rech.rpg.entity.Player;
+
 public class FileManager{
 
 	/*
