+++ src/com/rech/rpg/item/Inventory.java
@@ -6,7 +6,7 @@ import java.util.Scanner;
 
 import com.rech.rpg.Main;
 import com.rech.rpg.Menu;
-import com.rech.rpg.Player;
+import com.rech.rpg.entity.Player;
 
 public class Inventory{
 	//inventory stored as an array list which is just an array that can dynamically change its size, just a place holder for now
