+++ src/com/rech/rpg/map/event/StatusEffectEvent.java
@@ -3,7 +3,7 @@ package com.rech.rpg.map.event;
 import java.util.Scanner;
 
 import com.rech.rpg.Menu;
-import com.rech.rpg.Player;
+import com.rech.rpg.entity.Player;
 
 public class StatusEffectEvent extends Event{
 	protected enum statusEffect{
