+++ src/com/rech/rpg/map/location/Location.java
@@ -3,7 +3,7 @@ package com.rech.rpg.map.location;
 import java.util.Scanner;
 
 import com.rech.rpg.Menu;
-import com.rech.rpg.Player;
+import com.rech.rpg.entity.Player;
 import com.rech.rpg.map.Map.Direction;
 
 /**
