+++ src/com/rech/rpg/map/shop/Shop.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 import com.rech.rpg.Main;
 import com.rech.rpg.Menu;
-import com.rech.rpg.Player;
+import com.rech.rpg.entity.Player;
 import com.rech.rpg.item.Item;
 
 public class Shop {
